Scale websocket connection timeout with jobs
Updates GPU tests that use a heartbeat mechanism to scale the websocket
connection setup timeout based on the number of parallel test jobs. The
current timeout is fine in the vast majority of cases, but it appears to
be possible to get flaky timeouts when running 4 parallel jobs due to
all 4 Chrome instances starting up simultaneously and slowing down the
system enough.

Bug: 344009517
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5768094
Commit-Queue: Brian Sheedy <bsheedy@chromium.org>
Commit-Queue: Kenneth Russell <kbr@chromium.org>
Auto-Submit: Brian Sheedy <bsheedy@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1338689}

+++ content/test/gpu/gpu_tests/util/websocket_server.py
@@ -64,7 +64,7 @@ class WebsocketServer():
     self.websocket = None
     self.connection_received_event.clear()
 
-  def WaitForConnection(self, timeout: Optional[int] = None) -> None:
+  def WaitForConnection(self, timeout: Optional[float] = None) -> None:
     if self.websocket:
       return
     timeout = timeout or WEBSOCKET_SETUP_TIMEOUT_SECONDS
